Исправление ошибки преобразования строки в число(текущее время).

+++ src/variables.h
@@ -1,11 +1,11 @@
 #include <Arduino.h>
 /*Время включения и выключения преобразованные в формат целых чисел*/
-uint16_t Relay1Timeon=120000;
-uint16_t Relay1Timeoff=220000;
+uint32_t Relay1Timeon=120000;
+uint32_t Relay1Timeoff=220000;
 
 /*Время включения и выключения преобразованные в формат целых чисел*/
-uint16_t Relay2Timeon=193000;
-uint16_t Relay2Timeoff=183000;
+uint32_t Relay2Timeon=193000;
+uint32_t Relay2Timeoff=183000;
 
 /*Пины подключения релейных модулей*/
 const uint8_t Relay1Pin=3;
